:recycle: move function to progress service
